requirements.txt
python-json-logger==2.0.4
pytz==2023.3
tzdata==2023.3
asgiref==3.6.0
boto3==1.26.137
botocore==1.29.137
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
defusedxml==0.7.1
Django==4.2.1
django-allauth==0.54.0
django-environ==0.10.0
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
exceptiongroup==1.1.1
idna==3.4
iniconfig==2.0.0
jmespath==1.0.1
oauthlib==3.2.2
packaging==23.1
Pillow==9.5.0
pluggy==1.0.0
pycparser==2.21
PyJWT==2.6.0
pytest==7.3.1